+++ CHANGELOG.md
@@ -29,6 +29,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [#1429] Merge Armature will now allow you to merge humanoid bones with PhysBones attached in certain cases.
   - Specifically, child humanoid bones (if there are any) must be excluded from all attached Physbones. 
 - [#1437] Create Toggle for Selection now creates submenus as necessary when multiple items are selected, and creates toggles as children.
+- [#1499] When an audio source is controlled by an Object Toggle, disable the audio source when animations are blocked
+  to avoid it unintentionally being constantly active.
 
 ### Removed